+++ lib/actions/shopSyrattachPluginAttachments.action.php
@@ -1,15 +1,12 @@
 <?php
 /**
  * @author Serge Rodovnichenko <serge@syrnik.com>
- * @copyright Serge Rodovnichenko, 2021
- * @license Webasyst
+ * @copyright (c) 2014-2026, Serge Rodovnichenko
+ * @license http://www.webasyst.com/terms/#eula Webasyst
  */
 
 declare(strict_types=1);
 
-/**
- * Class shopSyrattachPluginAttachmentsAction
- */
 class shopSyrattachPluginAttachmentsAction extends waViewAction
 {
     /** @var shopSyrattachPlugin */
@@ -18,7 +15,7 @@ class shopSyrattachPluginAttachmentsAction extends waViewAction
     /**
      * @throws waException
      */
-    protected function preExecute()
+    protected function preExecute(): void
     {
         parent::preExecute();
         $this->plugin = wa('shop')->getPlugin('syrattach');
@@ -27,23 +24,22 @@ class shopSyrattachPluginAttachmentsAction extends waViewAction
     /**
      * @throws waException
      */
-    public function execute()
+    public function execute(): void
     {
-        if (!($product_id = waRequest::get('id', null, waRequest::TYPE_INT)))
+        $product_id = waRequest::get('id', null, waRequest::TYPE_INT);
+        if (!$product_id) {
             throw new waException(_wp('Product ID required'));
+        }
 
-        $product = new shopProduct($product_id);
-
-        $attachments = (new shopSyrattachFileModel())->getByProductId($product_id);
-        array_walk($attachments, function (&$attachment) {
-            $attachment['url'] = shopSyrattachPlugin::getFileUrl($attachment);
-        });
+        $product     = new shopProduct($product_id);
+        $file_model  = new shopSyrattachFileModel();
+        $attachments = $file_model->getByEntity('product', $product_id, true);
 
         $this->view->assign([
             'attachments'   => $attachments,
             'count'         => count($attachments),
             'max_file_size' => (int)waRequest::getUploadMaxFilesize(),
-            'product'       => $product
+            'product'       => $product,
         ]);
     }
 }